Add some missing cgroup test require and refactor the require check.
