+++ integration-cli/requirements_unix.go
@@ -3,68 +3,74 @@
 package main
 
 import (
-	"io/ioutil"
-	"path"
-
-	"github.com/opencontainers/runc/libcontainer/cgroups"
+	"github.com/docker/docker/pkg/sysinfo"
 )
 
 var (
+	// SysInfo stores information about which features a kernel supports.
+	SysInfo      *sysinfo.SysInfo
 	cpuCfsPeriod = testRequirement{
 		func() bool {
-			cgroupCPUMountpoint, err := cgroups.FindCgroupMountpoint("cpu")
-			if err != nil {
-				return false
-			}
-			if _, err := ioutil.ReadFile(path.Join(cgroupCPUMountpoint, "cpu.cfs_period_us")); err != nil {
-				return false
-			}
-			return true
+			return SysInfo.CPUCfsPeriod
 		},
 		"Test requires an environment that supports cgroup cfs period.",
 	}
 	cpuCfsQuota = testRequirement{
 		func() bool {
-			cgroupCPUMountpoint, err := cgroups.FindCgroupMountpoint("cpu")
-			if err != nil {
-				return false
-			}
-			if _, err := ioutil.ReadFile(path.Join(cgroupCPUMountpoint, "cpu.cfs_quota_us")); err != nil {
-				return false
-			}
-			return true
+			return SysInfo.CPUCfsQuota
 		},
 		"Test requires an environment that supports cgroup cfs quota.",
 	}
+	cpuShare = testRequirement{
+		func() bool {
+			return SysInfo.CPUShares
+		},
+		"Test requires an environment that supports cgroup cpu shares.",
+	}
 	oomControl = testRequirement{
 		func() bool {
-			cgroupMemoryMountpoint, err := cgroups.FindCgroupMountpoint("memory")
-			if err != nil {
-				return false
-			}
-			if _, err := ioutil.ReadFile(path.Join(cgroupMemoryMountpoint, "memory.memsw.limit_in_bytes")); err != nil {
-				return false
-			}
-
-			if _, err = ioutil.ReadFile(path.Join(cgroupMemoryMountpoint, "memory.oom_control")); err != nil {
-				return false
-			}
-			return true
-
+			return SysInfo.OomKillDisable
 		},
 		"Test requires Oom control enabled.",
 	}
 	kernelMemorySupport = testRequirement{
 		func() bool {
-			cgroupMemoryMountpoint, err := cgroups.FindCgroupMountpoint("memory")
-			if err != nil {
-				return false
-			}
-			if _, err := ioutil.ReadFile(path.Join(cgroupMemoryMountpoint, "memory.kmem.limit_in_bytes")); err != nil {
-				return false
-			}
-			return true
+			return SysInfo.KernelMemory
 		},
 		"Test requires an environment that supports cgroup kernel memory.",
 	}
+	memoryLimitSupport = testRequirement{
+		func() bool {
+			return SysInfo.MemoryLimit
+		},
+		"Test requires an environment that supports cgroup memory limit.",
+	}
+	swapMemorySupport = testRequirement{
+		func() bool {
+			return SysInfo.SwapLimit
+		},
+		"Test requires an environment that supports cgroup swap memory limit.",
+	}
+	memorySwappinessSupport = testRequirement{
+		func() bool {
+			return SysInfo.MemorySwappiness
+		},
+		"Test requires an environment that supports cgroup memory swappiness.",
+	}
+	blkioWeight = testRequirement{
+		func() bool {
+			return SysInfo.BlkioWeight
+		},
+		"Test requires an environment that supports blkio weight.",
+	}
+	cgroupCpuset = testRequirement{
+		func() bool {
+			return SysInfo.Cpuset
+		},
+		"Test requires an environment that supports cgroup cpuset.",
+	}
 )
+
+func init() {
+	SysInfo = sysinfo.New(true)
+}
